Restore setting controller fade
Removes help button in each section, which is jumpy. Fade is nicer though.
