+++ app/src/main/java/eu/kanade/tachiyomi/ui/setting/SettingsBackupController.kt
@@ -30,7 +30,7 @@ import java.io.File
 import java.util.concurrent.TimeUnit
 import eu.kanade.tachiyomi.data.preference.PreferenceKeys as Keys
 
-class SettingsBackupController : BaseSettingsController() {
+class SettingsBackupController : SettingsController() {
 
     /**
      * Flags containing information of what to backup.
