+++ app/src/main/java/eu/kanade/tachiyomi/ui/setting/SettingsDownloadController.kt
@@ -24,7 +24,7 @@ import uy.kohesive.injekt.injectLazy
 import java.io.File
 import eu.kanade.tachiyomi.data.preference.PreferenceKeys as Keys
 
-class SettingsDownloadController : BaseSettingsController() {
+class SettingsDownloadController : SettingsController() {
 
     private val db: DatabaseHelper by injectLazy()
