+++ app/src/main/java/eu/kanade/tachiyomi/ui/setting/SettingsLibraryController.kt
@@ -22,7 +22,7 @@ import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
 import eu.kanade.tachiyomi.data.preference.PreferenceKeys as Keys
 
-class SettingsLibraryController : BaseSettingsController() {
+class SettingsLibraryController : SettingsController() {
 
     private val db: DatabaseHelper = Injekt.get()
