+++ app/src/main/java/eu/kanade/tachiyomi/ui/setting/SettingsMainController.kt
@@ -1,11 +1,31 @@
 package eu.kanade.tachiyomi.ui.setting
 
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
 import androidx.preference.PreferenceScreen
-import com.bluelinelabs.conductor.RouterTransaction
 import eu.kanade.tachiyomi.R
+import eu.kanade.tachiyomi.ui.base.controller.withFadeTransaction
 import eu.kanade.tachiyomi.util.system.getResourceColor
+import eu.kanade.tachiyomi.util.system.openInBrowser
 
-class SettingsMainController : BaseSettingsController() {
+class SettingsMainController : SettingsController() {
+
+    init {
+        setHasOptionsMenu(true)
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        inflater.inflate(R.menu.settings, menu)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+            R.id.action_help -> activity?.openInBrowser(URL_HELP)
+        }
+
+        return super.onOptionsItemSelected(item)
+    }
 
     override fun setupPreferenceScreen(screen: PreferenceScreen) = with(screen) {
         titleRes = R.string.label_settings
@@ -81,6 +101,10 @@ class SettingsMainController : BaseSettingsController() {
     }
 
     private fun navigateTo(controller: SettingsController) {
-        router.pushController(RouterTransaction.with(controller))
+        router.pushController(controller.withFadeTransaction())
+    }
+
+    companion object {
+        private const val URL_HELP = "https://tachiyomi.org/help/"
     }
 }
