+++ app/src/main/java/eu/kanade/tachiyomi/ui/setting/SettingsReaderController.kt
@@ -5,7 +5,7 @@ import androidx.preference.PreferenceScreen
 import eu.kanade.tachiyomi.R
 import eu.kanade.tachiyomi.data.preference.PreferenceKeys as Keys
 
-class SettingsReaderController : BaseSettingsController() {
+class SettingsReaderController : SettingsController() {
 
     override fun setupPreferenceScreen(screen: PreferenceScreen) = with(screen) {
         titleRes = R.string.pref_category_reader
